% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_itr.R
\name{iterative_solve}
\alias{iterative_solve}
\title{Solving linear equations using iterative methods}
\usage{
iterative_solve(
  A,
  b,
  x0 = rep(0, length(b)),
  method = c("jacobi", "gauss-seidel"),
  max_iter = 1000L,
  d1 = 1e-10,
  d2 = 1e-10
)
}
\arguments{
\item{A}{A square matrix of coefficents, numeric without any NA or NaN}

\item{b}{A numeric vector being the right side of the equation}

\item{x0}{A numeric starting vector (default is a vector of zeros).}

\item{method}{A character string specifying the iterative method. Must be one of 
\code{"jacobi"} (default) or \code{"gauss-seidel"}.}

\item{max_iter}{An integer specifying the maximum number of iterations allowed (default: 1000)}

\item{d1}{A numeric value specifying the relative error tolerance (default: 1e-10)}

\item{d2}{A numeric value specifying the absolute error tolerance (default: 1e-10)}
}
\value{
A list containing two elements:
\item{solution}{A numeric solution vector x.}
\item{iterations}{The number of iterations performed, an integer.}
}
\description{
The function solves linear equations in form of Ax=b by using Jacobi or Gauss-Seidel methods (depending on choice) using Gill's stopping criteria, implemented in C language
}
\examples{
A <- matrix(c(10, -1, 2, -1, 11, -1, 2, -1, 10), nrow = 3, byrow = TRUE)
b <- c(6, 25, -11)
iterative_solve(A, b, method = "jacobi")
iterative_solve(A, b, x0 = c(1, 1, 1), method = "gauss-seidel")
}
\references{
"Notatki do wykładu Metody Numeryczne" by dr inż. Iwona Wróbel
}
