% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_lu.R
\name{lu_solve}
\alias{lu_solve}
\title{Solving linear equations using LU decomposition (Doolittle type)}
\usage{
lu_solve(A, b)
}
\arguments{
\item{A}{A square matrix of coefficents, numeric without any NA or NaN}

\item{b}{A numeric vector}
}
\value{
A numeric vector representing a solution
}
\description{
The function solves linear equations in form of Ax=b by using LU
decomposition with partial pivoting (GEPP) all implemented in C language.
}
\examples{
A <- matrix(c(2, 1, -1, -3, -1, 2, -2, 1, 2), nrow = 3, byrow = TRUE)
b <- c(8, -11, -3)
lu_solve(A, b)
}
\references{
"Notatki do wykładu Metody Numeryczne" by dr inż. Iwona Wróbel
}
